Finished saving data into db, retrieve & show data from db.
To do: Get current location, retrieve & show logs, delete a log

+++ OpenWIFI/src/main/java/classes/ApiModel.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ApiModel {
     private Wifi wifi;
-    public double dist;
+    public float dist;
 }
